+++ weblate/trans/models/conf.py
@@ -61,6 +61,9 @@ class WeblateConf(AppConf):
     # Enable lazy commits
     COMMIT_PENDING_HOURS = 24
 
+    # Automatically update vcs repositories daily
+    AUTO_UPDATE = False
+
     # List of quality checks
     CHECK_LIST = (
         'weblate.checks.same.SameCheck',