+++ docs/changes.rst
@@ -21,6 +21,7 @@ Released on ? 2018.
 * Removed support for custom commit messages.
 * Configurable editing mode (zen/full).
 * Added support for error reporting to Sentry.
+* Added support for automated daily update of repositories.
 
 weblate 3.1.1
 -------------